requirements.txt
numpy
pandas
sentence-transformers
huggingface_hub
umap-learn
hdbscan
scikit-learn
streamlit
plotly
